Working with Heroku endpoints
