+++ invoice.json
@@ -5,10 +5,10 @@
         "credits_total": 0,
         "id": "22e1c650-291e-48b7-b462-4bbd9eba4a0e",
         "number": 62983158,
-        "period_end": "2022-02-22",
+        "period_end": "2022-02-23",
         "period_start": "2022-02-01",
         "state": 0,
         "total": 0,
-        "updated_at": "2022-02-22T02:45:31Z"
+        "updated_at": "2022-02-23T02:44:02Z"
     }
 ]
