--- src/GithubAccount.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Ishifoev\HerokuApi;
-
-class GithubAccount {
-    public function getGithubAccount() {
-        var_dump('github');
-    }
-}
